third_party/WebKit/LayoutTests/fast/dynamic/move-node-with-selection.html

<!doctype html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../editing/assert_selection.js"></script>
<script>
test(() => assert_selection(
    '<div id="dest"></div><div id="target">Lor^em i|psum</div>',
    selection => {
        var target = selection.document.getElementById('target');
        var dest = selection.document.getElementById('dest');
        target.style.display = 'inline';
        dest.appendChild(target);
    },
    [
        '<div id="dest">',
            '<div id="target" style="display: inline;">Lorem ipsum</div>',
        '</div>|',
    ].join('')),
    'Modify and move a container node of selection');
</script>
